ml/requirements.txt
flask==3.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
mysql-connector-python==8.4.0